code-refinement-1

+++ src/main/java/com/ecommerce/service/ProductService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import com.ecommerce.dto.CategoryDto;
 import com.ecommerce.dto.ProductDto;
+import com.ecommerce.exceptions.ItemNotFoundException;
 import com.ecommerce.model.Product;
 import com.ecommerce.repository.ProductRepository;
 
@@ -17,7 +18,7 @@ public class ProductService {
 	private ProductRepository productRepository;
 	
 //	CREATE
-	public Product addProduct(Product product) {
+	public Product createProduct(Product product) {
 		return productRepository.save(product);
 	}
 	
@@ -26,15 +27,14 @@ public class ProductService {
 		return productRepository.findAll();
 	}
 	
+//	this will fetch details lazily excluding category fields
 	public Product getById(Integer id) {
-		return productRepository.findById(id).get();
+		return productRepository.findById(id)
+				.orElseThrow(()-> new ItemNotFoundException("No item found with the given description"));
 	}
 	
-	public List<Product> getProductByName(String name) {
-		return productRepository.findByNameContaining(name);
-	}
-	
-	public ProductDto findProductById(Integer id) {
+//	this will fetch details eagerly including category field using dtos
+	public ProductDto getByIdEagerFetch(Integer id) {
 
 		Product product = productRepository.findById(id).get();
 
@@ -54,8 +54,15 @@ public class ProductService {
 
 	}
 	
+	public Product getByName(String name) {
+		if(productRepository.existsByName(name)) {
+			return productRepository.findByName(name);
+		}
+		throw new ItemNotFoundException("No item found with the given description");
+	}
+	
 //	UPDATE
-	public List<Product> updateProduct(int id, Product product) {
+	public Product updateById(int id, Product product) {
 		if (productRepository.existsById(id)) {
 
 			Product product2 = productRepository.findById(id).get();
@@ -64,15 +71,21 @@ public class ProductService {
 			product2.setRating(product.getRating());
 			product2.setPrice(product.getPrice());
 
-			productRepository.save(product2);
+			return productRepository.save(product2);
 		}
-		return productRepository.findAll();
+		
+		throw new ItemNotFoundException("No item found with the given description");
 	}
 	
 //	DELETE
-	public List<Product> deleteProduct(int id) {
-		productRepository.deleteById(id);
-		return productRepository.findAll();
+	public List<Product> deleteById(int id) {
+		if (productRepository.existsById(id)) {
+			productRepository.deleteById(id);
+			return productRepository.findAll();
+		}
+		
+		throw new ItemNotFoundException("No item found with the given description");
+		
 	}
 
 }
